move tests out of folders for build files
