+++ packages/owlbot-bootstrapper/test/mono-repo.test.ts
@@ -14,13 +14,13 @@
 
 import {describe, it} from 'mocha';
 import {Octokit} from '@octokit/rest';
-import {Language} from '../../common-container/interfaces';
+import {Language} from '../common-container/interfaces';
 import nock from 'nock';
 import {execSync} from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs';
-import {MonoRepo} from '../../common-container/mono-repo';
-import * as utils from '../../common-container/utils';
+import {MonoRepo} from '../common-container/mono-repo';
+import * as utils from '../common-container/utils';
 import assert from 'assert';
 
 nock.disableNetConnect();
